# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "這是一個使用 Gin 框架構建的 CRUD API 服務，包含用戶管理、書籍管理、產品管理、文章管理、認證、限流器、工作池等功能。",
        "title": "Go Gin CRUD API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/worker-pool/batch-submit": {
            "post": {
                "description": "向工作池批量提交多個任務",
                "tags": [
                    "worker-pool"
                ],
                "summary": "批量提交任務",
                "parameters": [
                    {
                        "description": "批量提交任務請求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.BatchSubmitTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.BatchSubmitTaskResponse"
                        }
                    }
                }
            }
        },
        "/api/worker-pool/create": {
            "post": {
                "description": "創建一個新的工作池，用於處理任務",
                "tags": [
                    "worker-pool"
                ],
                "summary": "創建工作池",
                "parameters": [
                    {
                        "description": "創建工作池請求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.CreateWorkerPoolRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.WorkerPoolStatus"
                        }
                    }
                }
            }
        },
...
        "/api/worker-pool/result": {
            "get": {
                "description": "獲取指定任務的執行結果",
                "tags": [
                    "worker-pool"
                ],
                "summary": "獲取任務結果",
                "parameters": [
                    {
                        "type": "string",
                        "description": "工作池 ID",
                        "name": "pool_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "任務 ID",
                        "name": "task_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.WorkerPoolResult"
                        }
                    }
                }
            }
        },
        "/api/worker-pool/shutdown": {
            "post": {
                "description": "優雅關閉指定工作池，等待所有任務完成",
                "tags": [
                    "worker-pool"
                ],
                "summary": "優雅關閉工作池",
                "parameters": [
                    {
                        "type": "string",
                        "description": "工作池 ID",
                        "name": "pool_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "超時時間（秒，預設 30 秒）",
                        "name": "timeout",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/worker-pool/status": {
            "get": {
                "description": "獲取指定工作池的狀態資訊",
                "tags": [
                    "worker-pool"
                ],
                "summary": "獲取工作池狀態",
                "parameters": [
                    {
                        "type": "string",
                        "description": "工作池 ID",
                        "name": "pool_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.WorkerPoolStatus"
                        }
                    }
                }
            }
        },
        "/api/worker-pool/submit": {
            "post": {
                "description": "向工作池提交一個任務",
                "tags": [
                    "worker-pool"
                ],
                "summary": "提交任務",
                "parameters": [
                    {
                        "description": "提交任務請求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.SubmitTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.SubmitTaskResponse"
                        }
                    }
                }
            }
        },
...
        "/stream/chunked": {
            "get": {
                "description": "以 chunked 編碼逐筆回傳 JSON 行（NDJSON），可用 Query count 指定筆數（預設 10）",
                "produces": [
                    "application/x-ndjson"
                ],
                "tags": [
                    "streaming"
                ],
                "summary": "Chunked 串流",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "輸出筆數",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "application/x-ndjson",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/stream/sse": {
            "get": {
                "description": "Server-Sent Events 串流，每秒推送一筆事件，可用 Query seconds 指定秒數（預設 30）",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "streaming"
                ],
                "summary": "SSE 串流",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "串流持續秒數",
                        "name": "seconds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "text/event-stream",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "go-gin-crud_internal_dto.BatchSubmitTaskRequest": {
            "type": "object",
            "required": [
                "pool_id",
                "tasks"
            ],
            "properties": {
                "pool_id": {
                    "description": "工作池 ID",
                    "type": "string"
                },
                "tasks": {
                    "description": "任務列表",
                    "type": "array",
                    "maxItems": 100,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/go-gin-crud_internal_dto.SubmitTaskRequest"
                    }
                }
            }
        },
        "go-gin-crud_internal_dto.BatchSubmitTaskResponse": {
            "type": "object",
            "properties": {
                "pool_id": {
                    "type": "string"
                },
                "rejected": {
                    "description": "被拒絕的任務數",
                    "type": "integer"
                },
                "results": {
                    "description": "每個任務的提交結果",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-gin-crud_internal_dto.SubmitTaskResponse"
                    }
                },
                "submitted": {
                    "description": "成功提交的任務數",
                    "type": "integer"
                },
                "submitted_at": {
                    "type": "string"
                },
                "total_tasks": {
                    "type": "integer"
                }
            }
        },
...
        "go-gin-crud_internal_dto.CreateWorkerPoolRequest": {
            "type": "object",
            "required": [
                "pool_id",
                "worker_count"
            ],
            "properties": {
                "pool_id": {
                    "description": "工作池 ID",
                    "type": "string"
                },
                "queue_size": {
                    "description": "任務佇列大小（可選，預設無限制）",
                    "type": "integer",
                    "minimum": 1
                },
                "worker_count": {
                    "description": "Worker 數量",
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1
                }
            }
        },
...
        "go-gin-crud_internal_dto.SubmitTaskRequest": {
            "type": "object",
            "required": [
                "pool_id",
                "task_id",
                "type"
            ],
            "properties": {
                "data": {
                    "description": "任務數據",
                    "type": "object",
                    "additionalProperties": true
                },
                "pool_id": {
                    "description": "工作池 ID",
                    "type": "string"
                },
                "priority": {
                    "description": "優先級（預設 0）",
                    "type": "integer"
                },
                "task_id": {
                    "description": "任務 ID",
                    "type": "string"
                },
                "timeout": {
                    "description": "超時時間（秒，預設 30 秒）",
                    "type": "integer"
                },
                "type": {
                    "description": "任務類型",
                    "type": "string"
                }
            }
        },
        "go-gin-crud_internal_dto.SubmitTaskResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "pool_id": {
                    "type": "string"
                },
                "status": {
                    "description": "submitted, rejected",
                    "type": "string"
                },
                "submitted_at": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                }
            }
        },
...
        "go-gin-crud_internal_dto.WorkerPoolResult": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "完成時間",
                    "type": "string"
                },
                "duration": {
                    "description": "執行時間（毫秒）",
                    "type": "integer"
                },
                "error": {
                    "description": "錯誤訊息",
                    "type": "string"
                },
                "result": {
                    "description": "執行結果",
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "description": "狀態：success, failed, timeout, cancelled",
                    "type": "string"
                },
                "task_id": {
                    "description": "任務 ID",
                    "type": "string"
                },
                "worker_id": {
                    "description": "處理該任務的 Worker ID",
                    "type": "integer"
                }
            }
        },
        "go-gin-crud_internal_dto.WorkerPoolStatus": {
            "type": "object",
            "properties": {
                "active_workers": {
                    "description": "活躍的 Worker 數量",
                    "type": "integer"
                },
                "completed_tasks": {
                    "description": "已完成任務數",
                    "type": "integer"
                },
                "created_at": {
                    "description": "創建時間",
                    "type": "string"
                },
                "failed_tasks": {
                    "description": "失敗任務數",
                    "type": "integer"
                },
                "is_running": {
                    "description": "是否運行中",
                    "type": "boolean"
                },
                "pool_id": {
                    "description": "工作池 ID",
                    "type": "string"
                },
                "queue_capacity": {
                    "description": "佇列容量（-1 表示無限制）",
                    "type": "integer"
                },
                "queue_size": {
                    "description": "當前佇列大小",
                    "type": "integer"
                },
                "total_tasks": {
                    "description": "總任務數",
                    "type": "integer"
                },
                "worker_count": {
                    "description": "Worker 數量",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "請在值前加上 \"Bearer \" 前綴，例如：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}